requirements.txt
openai
pinecone-client
pinecone
sentence-transformers
numpy
python-dotenv
